% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_squid.R
\name{theme_squid}
\alias{theme_squid}
\title{ggplot theme for minimalist plotting}
\usage{
theme_squid(base_size = 12)
}
\arguments{
\item{base_size}{base font size, given in pts}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
Theme used for plotting SQUID component plots. Modification of theme_classic from original ggplot2 package, with consistent font sizing across text elements and with black text, rather than dark grey.
}
\examples{
library("ggplot2")

ggplot(mtcars) +
  aes(x = wt, y = mpg, colour = factor(gear)) +
  geom_point() +
  theme_squid()
}
